Use the long long data type to handle large integers that exceed the range of int: This ensures the program won’t break due to overflow when dealing with very large integers

+++ ConditionalStructure/34_CompareFractions.cpp
@@ -27,7 +27,7 @@ using namespace std;
 
 int main()
 {
-	int a,b,c,d,crossMult1,crossMult2;
+	long long a,b,c,d,crossMult1,crossMult2;
 	cout<<"Input a non-zero integer, a: ";
 	cin>>a;
 	cout<<"Input a non-zero integer, b: ";
